Fix warning

+++ wmr100.c
@@ -143,7 +143,7 @@ int wmr_send_packet_ready(WMR *wmr) {
 }
 
 void wmr_print_state(WMR *wmr) {
-    fprintf(stderr, "WMR: HID: %08x\n", wmr->hid);
+  fprintf(stderr, "WMR: HID: %08x\n", (unsigned int)wmr->hid);
 }
 
 int wmr_close(WMR *wmr) {
